requirements.txt
fastapi==0.115.6
ffmpeg-python==0.2.0
pydantic==2.10.4
requests==2.32.3
uvicorn==0.34.0
python-multipart
python-dotenv==1.0.1
yt-dlp==2023.07.06
runpod
httpx
youtube-transcript-api==0.6.3

celery==5.3.4
redis==4.6.0